+++ data/init-data.ts
@@ -1,0 +1,76 @@
+import {
+  ItemsRes,
+  backpacks,
+  helmets,
+  guns,
+  armors,
+  Item,
+  Items,
+  rigs,
+} from "./item-data";
+import axios from "axios";
+
+const BASE_URL = "https://api.tarkov.dev/graphql";
+
+// 初回データ取得
+export const initData = async () => {
+  fetchItems();
+};
+
+const fetchItems = async () => {
+  // query Body
+  const query = `
+      {
+        items {
+          name
+          types
+        }
+      }
+    `;
+  try {
+    const response = await axios.post(BASE_URL, {
+      query: query,
+    });
+    const res: ItemsRes = response.data.data;
+    if (res && res.items) {
+      Items.push(...res.items);
+      sortingItems(res.items);
+    }
+  } catch (error) {
+    console.error("Error fetching data:", error);
+  }
+};
+
+// アイテムをタイプ別に振り分け
+const sortingItems = (items: Item[]) => {
+  console.time("sortingItems");
+  items.forEach((item) => {
+    if (item.types.includes("gun")) {
+      guns.push(item);
+    } else if (item.types.includes("armor")) {
+      armors.push(item);
+    } else if (item.types.includes("helmet")) {
+      helmets.push(item);
+    } else if (item.types.includes("backpack")) {
+      backpacks.push(item);
+    } else if (item.types.includes("rig")) {
+      rigs.push(item);
+    }
+  });
+  console.timeEnd("sortingItems");
+};
+
+const fetchBosses = async () => {
+  const query = `
+    {
+    bosses{
+      name
+      imagePosterLink
+      imagePortraitLink
+      health{
+        max
+        bodyPart
+      }
+    }
+}`;
+};
