+++ data/item-data.ts
@@ -1,4 +1,4 @@
-import axios from "axios";
+import { maps } from "./map-data";
 
 export const ITEM_CATEGORIES = {
   armors: "armors",
@@ -9,16 +9,16 @@ export const ITEM_CATEGORIES = {
   rigs: "rigs",
 };
 
-interface Item {
+export interface Item {
   name: string;
   types: string[];
 }
 
-interface ItemsRes {
+export interface ItemsRes {
   items: Item[];
 }
 
-interface equipment{
+export interface equipment{
   gun: string;
   armor: string;
   helmet: string;
@@ -26,26 +26,13 @@ interface equipment{
   rig: string;
 }
 
-const Items: Item[] = [];
+export const Items: Item[] = [];
 
-const guns: Item[] = [];
-const armors: Item[] = [];
-const helmets: Item[] = [];
-const backpacks: Item[] = [];
-const rigs: Item[] = [];
-
-const maps = [
-  "Customs",
-  "Factory",
-  "Interchange",
-  "LightHouse",
-  "Labs",
-  "Reserve",
-  "Shoreline",
-  "Streets",
-  "The Lab",
-  "Woods",
-];
+export const guns: Item[] = [];
+export const armors: Item[] = [];
+export const helmets: Item[] = [];
+export const backpacks: Item[] = [];
+export const rigs: Item[] = [];
 
 export const getItems = (itemType: string) => {
   switch (itemType) {
@@ -93,50 +80,3 @@ export const returnEquipment = () => {
   }
   return equipment;
 }
-
-// 初回データ取得
-export const initData = async () => {
-  const BASE_URL = "https://api.tarkov.dev/graphql";
-
-  // query Body
-  const query = `
-    {
-      items {
-        name
-        types
-      }
-    }
-  `;
-
-  try {
-    const response = await axios.post(BASE_URL, {
-      query: query,
-    });
-    const res: ItemsRes = response.data.data;
-    if (res && res.items) {
-      Items.push(...res.items);
-      sortingItems(res.items);
-    }
-  } catch (error) {
-    console.error("Error fetching data:", error);
-  }
-};
-
-// アイテムをタイプ別に振り分け
-const sortingItems = (items: Item[]) => {
-  console.time("sortingItems");
-  items.forEach((item) => {
-    if (item.types.includes("gun")) {
-      guns.push(item);
-    } else if (item.types.includes("armor")) {
-      armors.push(item);
-    } else if (item.types.includes("helmet")) {
-      helmets.push(item);
-    } else if (item.types.includes("backpack")) {
-      backpacks.push(item);
-    } else if (item.types.includes("rig")) {
-      rigs.push(item);
-    }
-  });
-  console.timeEnd("sortingItems");
-};
