+++ data/map-data.ts
@@ -1,0 +1,12 @@
+export const maps = [
+    "Customs",
+    "Factory",
+    "Interchange",
+    "LightHouse",
+    "Labs",
+    "Reserve",
+    "Shoreline",
+    "Streets",
+    "The Lab",
+    "Woods",
+  ];
